finally update

+++ server/model/RBAC.go
@@ -50,7 +50,7 @@ type Menu struct {
 	Component string `gorm:"type:varchar(50);comment:组件" json:"component"`
 	Icon      string `gorm:"type:varchar(50);comment:菜单图标" json:"icon"`
 	ParentId  int    `gorm:"comment:父菜单id" json:"parent_id"`
-	OrderNum  int    `gorm:"type:tinyint;default:0;comment:显示排序" json:"order_num"`
+	OrderNum  int8   `gorm:"type:tinyint;default:0;comment:显示排序" json:"order_num"`
 	IsHidden  int8   `gorm:"type:tinyint(1);default:0;comment:是否隐藏(0-否 1-是)" json:"is_hidden"`
 	KeepAlive int8   `gorm:"type:tinyint(1);default:1" json:"keep_alive"`
 	Redirect  string `gorm:"type:varchar(50);comment:跳转路径" json:"redirect"`
